requirements.txt
pandas
sentence-transformers
weaviate-client
langchain
pyspark
openpyxl
python-dotenv
numpy
scikit-learn
scipy
matplotlib
seaborn
nltk
beautifulsoup4
requests
tqdm
flask
gunicorn
pytest       
langchain_community
langchain_openai
langchain_weaviate
langchain_community
openpyxl